Add emojis to be hip with the kids
Wplacer (a botting tool I don't use but am aware of) uses emojis in almost every single log. I realized the chart emojis could be a good way to convey if a SEAL log is good or bad (decreasing or increasing) if you haven't seen an alert for that template in a while. So there you go: emojos.